requirements.txt
jinja2
pdfkit
requests
openpyxl
mysql-connector-python